Refactorizando codigo

+++ src/main/java/com/anthonyponte/productos/repository/ProductoRepository.java
@@ -6,6 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.anthonyponte.productos.model.Producto;
 
 @Repository
-public interface ProductoRepository extends JpaRepository<Producto, Integer> {
-
+public interface ProductoRepository extends JpaRepository<Producto, Long> {
 }
